requirements.txt
#core
asgiref==3.5.0
Django==4.0.3
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary==2.9.3

#djangorest
djangorestframework==3.13.1
django-filter==21.1

#sugar
django-environ==0.8.1

#requests
requests==2.27.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
urllib3==1.26.8

#dev API-docs
MarkupSafe==2.1.1
coreapi==2.3.3
coreschema==0.0.4
itypes==1.2.0
jinja2==3.0.3
uritemplate==4.1.1
importlib-metadata==4.11.3
markdown==3.3.6
zipp==3.7.0

#tests
Faker==13.3.2
factory-boy==3.2.1
python-dateutil==2.8.2
six==1.16.0
